+++ processing/subsea_cable_processing_provider.py
@@ -21,6 +21,7 @@ from .merge_mbes_rasters_algorithm import MergeMBESRastersAlgorithm
 
 from .import_ship_outline_algorithm import ImportShipOutlineAlgorithm
 from .place_ship_outlines_algorithm import PlaceShipOutlinesAlgorithm
+from .plot_line_segments_from_table_algorithm import PlotLineSegmentsFromTableAlgorithm
 
 
 class SubseaCableProcessingProvider(QgsProcessingProvider):
@@ -53,6 +54,8 @@ class SubseaCableProcessingProvider(QgsProcessingProvider):
         self.addAlgorithm(ImportShipOutlineAlgorithm())
         print('Registering PlaceShipOutlinesAlgorithm...')
         self.addAlgorithm(PlaceShipOutlinesAlgorithm())
+        print('Registering PlotLineSegmentsFromTableAlgorithm...')
+        self.addAlgorithm(PlotLineSegmentsFromTableAlgorithm())
 
     def id(self):
         """